PegOutRequest struct (#114)
* Modified PegOut struct to match contract implementation
* Removed fee from struct and its usage

+++ examples/union/participants/committee.rs
@@ -101,7 +101,7 @@ impl Committee {
         request_pegin_txid: Txid,
         amount: u64,
         accept_pegin_sighash: Vec<u8>,
-        slot_index: usize,
+        slot_index: u64,
         rootstock_address: String,
         reimbursement_pubkey: PublicKey,
     ) -> Result<()> {
